icinga2 (2.10.3-2~bpo9+1) stretch-backports; urgency=medium

  * Rebuild for stretch-backports.

icinga2 (2.10.3-2) unstable; urgency=medium

  * Team upload.
  * Add patch to fix comparepasswords_issafe test failure on i386.
  * Set CTEST_OUTPUT_ON_FAILURE environment variable.

icinga2 (2.10.3-1) unstable; urgency=medium

  * Team upload.
  * New upstream release.
  * Bump Standards-Version to 4.3.0, no changes.
  * Use .maintscript files instead of dpkg-maintscript-helper directly.
  * Add lintian overrides for file-references-package-build-path.

icinga2 (2.10.2-1~bpo9+1) stretch-backports; urgency=medium

  * Rebuild for stretch-backports.

icinga2 (2.10.2-1) unstable; urgency=medium

  * Team upload.
  * New upstream release.
  * Drop 42_mips_boost_bind_workaround patch, applied upstream.

icinga2 (2.10.1-3~bpo9+1) stretch-backports; urgency=medium

  * Rebuild for stretch-backports.

icinga2 (2.10.1-3) unstable; urgency=medium

  * Team upload.
  * Move init script & systemd service from -common to -bin package.
    (closes: #912599)

icinga2 (2.10.1-2~bpo9+1) stretch-backports; urgency=medium

  * Rebuild for stretch-backports.

icinga2 (2.10.1-2) unstable; urgency=medium

  * Team upload.
  * Enable USE_SYSTEMD option & change icinga2.service to notify type.
    (closes: #911672)

icinga2 (2.10.1-1) unstable; urgency=medium

  * Team upload.
  * New upstream release.
  * Update www.icinga.com URLs to icinga.com, and dev.icinga.com to GitHub.
  * Drop spelling-errors.patch, applied upstream.
  * Move from experimental to unstable.

icinga2 (2.10.0-1~exp1) experimental; urgency=medium

  * Team upload.
  * New upstream release.
  * Add patch to fix spelling errors.

icinga2 (2.9.2-1~bpo9+1) stretch-backports; urgency=medium

  * Rebuild for stretch-backports.
  * Add myself to Uploaders for this backport.
  * Update the branch name in Vcs-Git field and gbp.conf.

icinga2 (2.9.2-1) unstable; urgency=medium

  * Team upload.
  * New upstream release.
  * Bump Standards-Version to 4.2.1, no changes.

icinga2 (2.9.1-1) unstable; urgency=medium

  * Team upload.
  * New upstream release.
  * Drop spelling-errors.patch, applied upstream.

icinga2 (2.9.0-1) unstable; urgency=medium

  * Team upload.
  * New upstream release.
  * Add patch by Jens Holzkämper to use check_postgres path as
    provided by check-postgres package.
    (closes: #898833)
  * Apply patch by Dmitry Bogatov to also compress 01-about.md.
    (closes: #898701)
  * Bump Standards-Version to 4.1.5, no changes.
  * Update license & copyright for mail notification scripts.
  * Bump minimum required cmake version to 2.8.8.
  * Refresh patches.
  * Drop icinga2-studio package, removed upstream.
  * Update ido install files for shim libraries.
  * Drop libicinga2 package, libraries no longer built.
  * Update spelling-errors.patch for "doesn't" typo.

icinga2 (2.8.4-2) unstable; urgency=medium

  * Team upload.
  * Fix version autopkgtest.

icinga2 (2.8.4-1) unstable; urgency=medium

  * Team upload.
  * Fix jQuery usage in icinga2-classicui.
    (closes: #867884, #875686)
  * Update Swedish debconf translation. Thanks to Jonatan Nyberg.
    (closes: #873009)
  * Change libreadline6 Recommends to libreadline7.
    (closes: #880488)
  * Add upstream metadata.
  * Move from experimental to unstable.

icinga2 (2.8.4-1~exp1) experimental; urgency=medium

  * Team upload.

  [ Bas Couwenberg ]
  * New upstream release.
    - CVEs fixes in 2.8.2:
      CVE-2017-16933, CVE-2018-6532, CVE-2018-6533,
      CVE-2018-6534,  CVE-2018-6535, CVE-2018-6536
    (closes: #897301, #891333, #883247)
  * Add gbp.conf to use pristine-tar by default.
  * Update copyright file, changes:
    - Update copyright years for Icinga Development Team.
    - Use stand-alone license paragraphs
    - Add license & copyright for socketpair
  * Restructure control file with cme.
  * Change priority from extra to optional.
  * Sort (build) dependencies.
  * Drop obsolete dbg package.
  * Update Vcs-* URLs for Salsa.
  * Update copyright format URL to use HTTPS.
  * Update various debian.org & icinga.com URLs to use HTTPS.
  * Bump Standards-Version to 4.1.4, changes: priority, copyright format.
  * Enable parallel globally.
  * Simplify 'disable unity builds' rules.
  * Drop obsolete dh-systemd build dependency.
  * Use DEB_BUILD_ARCH instead of DEB_HOST_ARCH, and not unconditionally.
  * Drop obsolete mysql (build) dependency alternatives.
  * Strip trailing whitespace from changelog.
  * Enable all hardening buildflags.
  * Add patch to fix spelling errors.
  * Move spelling-error-in-binary override from icinga2-bin to libicinga2.
  * Drop unused overrides for apache2-deprecated-auth-config.
  * Move rm_conffile from prerm to postrm.
  * Update systemd service file, changes:
    - Remove obsolete syslog.target
    - Add Documentation key
  * Add lintian overrides for hardening-no-fortify-functions.
  * Override dh_missing to use --list-missing.
  * Sort rules in order of execution.
  * Reorder and align install files.
  * Install IDO SQL files from debian/tmp.
  * Explicitly remove files not included in any package.

  [ Dimitri John Ledkov ]
  * Make sure icinga2-common has /etc/icinga2/pki folder, which appears to
    still be in use.

icinga2 (2.7.0-1) unstable; urgency=medium

  * [3f2f6e0] New upstream version 2.7.0
  * [8861bd7] Improve systemd unit file
  * [f25990d] fix linebreak
  * [98b7e3f] Update debhelper dependency (Closes: #854349)
  * [8c35696] Update swedish translation (Closes: #855364)
  * [6f58b41] Install nsca_api plugin

icinga2 (2.6.3-1) experimental; urgency=medium

  * [83b389e] New upstream version 2.6.3

icinga2 (2.6.2-1~exp1) experimental; urgency=medium

  * [16eb968] Depend on network.target and fix typo in service deps
  * [7eee479] New upstream version 2.6.2

icinga2 (2.6.1-1~exp1) experimental; urgency=medium

  * [8cc068c] New upstream version 2.6.1
  * [509d654] Allow mysqlclient-dev dependency to be backwards compatible
  * [8cc068c] New upstream version 2.6.1
  * [95aa4c2] Remove obsolete patch 41_ido_mysql57

icinga2 (2.6.0-2) unstable; urgency=medium

  * [e0f34e4] Add patch 41_ido_mysql57 - for compatibility with MySQL 5.7
  * [6512d11] Remove obsolete patch 30_use_libmysql (empty)
  * [16e341b] Recommend default-mysql-{server,client}
  * [749f1fa] Adding missing dependency on lsb-base
  * [4ed1e10] Update initscript to ignore SIGPIPE in daemon
  * [5c10b76] Add patch 42_mips_boost_bind_workaround
  * [86a23e5] makeshlibs: no scripts, we don't export shared libraries

icinga2 (2.6.0-1) unstable; urgency=medium

  [ Alexander Wirt ]
  * [4e44b28] New upstream version 2.6.0
    - Fix unittests (Closes: #838358)

  [ Gunnar Beutner ]
  * [de73edf] Add support for versioned private libraries

icinga2 (2.5.4-3) unstable; urgency=medium

  * [e09f7d4] Link against libmysqlclient (Closes: #825079)

icinga2 (2.5.4-2) unstable; urgency=medium

  * [f749f15] Disable unity builds for arm
  * [33face7] Disable unity builds for mips and mipsel
    (Closes: #836589)

icinga2 (2.5.4-1) unstable; urgency=medium

  * [7cf17d9] Recommend specific icinga-doc version
  * [d2c5d89] New upstream version 2.5.4

+++ lib/base/convert.cpp
@@ -1,6 +1,6 @@
 /******************************************************************************
  * Icinga 2                                                                   *
- * Copyright (C) 2012-2016 Icinga Development Team (https://www.icinga.org/)  *
+ * Copyright (C) 2012-2018 Icinga Development Team (https://icinga.com/)      *
  *                                                                            *
  * This program is free software; you can redistribute it and/or              *
  * modify it under the terms of the GNU General Public License                *
@@ -39,7 +39,7 @@ String Convert::ToString(double val)
 	double fractional = std::modf(val, &integral);
 
 	if (fractional == 0)
-		return Convert::ToString(static_cast<long>(val));
+		return Convert::ToString(static_cast<long long>(val));
 
 	std::ostringstream msgbuf;
 	msgbuf << std::fixed << val;
